clickscript/pyproject.toml
[tool.poetry]
name = "clickscript"
version = "0.1.0"
description = "A Python script that simulates mouse clicks at the current mouse position."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pynput = "^1.7.6"
pyobjc-framework-Quartz = "^11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"